requirements.txt
numpy==1.24.1
pillow==9.3.0
torch==2.6.0
torchvision==0.21.0
opencv-python==4.11.0.86
scikit-learn==1.6.1
pandas==2.2.3
tqdm==4.67.1
matplotlib==3.10.1
python-dateutil==2.8.2
typing_extensions==4.12.2
flask==2.3.3
gunicorn==21.2.0
fastapi==0.103.1
uvicorn==0.23.2
mlflow==2.9.0
pytest==7.4.3
python-multipart==0.0.9
prometheus-client==0.18.0
pyyaml==6.0.1
onnx==1.15.0
onnxruntime==1.16.3
# Additional requirements for audio processing
librosa>=0.10.0
soundfile>=0.12.1 